Align REST API router split with upstream
