+++ core/src/core/rest_api.py
@@ -1,4 +1,4 @@
-# Copyright 2024 DataRobot, Inc.
+# Copyright 2025 DataRobot, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,362 +11,97 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+"""Main FastAPI application factory for the Data Analyst API."""
+
 from __future__ import annotations
 
-import asyncio
-import io
-import json
 import os
-import sys
-import tempfile
-from typing import Any, Iterable, List, Union, cast
+from typing import Any
 
-import chardet
-import pandas as pd
-from fastapi import (
-    APIRouter,
-    BackgroundTasks,
-    Depends,
-    FastAPI,
-    Form,
-    HTTPException,
-    Request,
-    Response,
-    UploadFile,
-)
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.openapi.utils import get_openapi
-from fastapi.responses import StreamingResponse
-from openai.types.chat.chat_completion_message_param import ChatCompletionMessageParam
-from openai.types.chat.chat_completion_user_message_param import (
-    ChatCompletionUserMessageParam,
-)
-from openpyxl import Workbook
-from openpyxl.drawing.image import Image as XLImage
-from openpyxl.utils.dataframe import dataframe_to_rows
-from starlette.background import BackgroundTask
 from starlette.types import Lifespan
 
+import core.routers.chats as chats_routes
+import core.routers.database as database_routes
+import core.routers.datasets as datasets_routes
+import core.routers.dictionaries as dictionaries_routes
+import core.routers.external_data_stores as external_data_store_routes
+import core.routers.registry as registry_routes
+import core.routers.user as user_routes
+from core import deps, file_utils
 from core import middleware as core_middleware
-from core.analyst_db import (
-    AnalystDB,
-    DatasetMetadata,
-    DataSourceType,
-    ExternalDataStoreNameDataSourceType,
-    InternalDataSourceType,
-    get_data_source_type,
-)
-from core.api_exceptions import ExceptionBody
-from core.chat_dataset_helper import cleanup_message_datasets
 from core.data_analyst_telemetry import telemetry
-from core.database_helpers import NoDatabaseOperator, get_external_database
-from core.datarobot_client import get_visitors_token, use_user_token
-from core.datarobot_dataset_handler import DatasetSparkRecipe, DataSourceRecipe
 from core.logging_helper import get_logger
-
-sys.path.append(os.path.dirname(os.path.realpath(__file__)))
-
-from datarobot_genai.core.utils.token_tracking import (
-    HeuristicTokenCountingStrategy,
-    TokenUsageTracker,
-    count_messages_tokens,
-)
-
-from core.api import (
-    AnalysisGenerationError,
-    list_registry_datasets,
-    load_registry_datasets,
-    log_memory,
-    process_data_and_update_state,
-    register_remote_registry_datasets,
-    run_complete_analysis,
-    summarize_conversation,
-    sync_data_sources_and_datasets,
-)
-from core.constants import (
-    ALTERNATIVE_LLM_BIG,
-    CONTEXT_WARNING_THRESHOLD,
-    MODEL_CONTEXT_WINDOW,
-)
-from core.schema import (
-    AnalystChatMessage,
-    AnalystDataset,
-    ChatCreate,
-    ChatMessagePayload,
-    ChatRequest,
-    ChatResponse,
-    ChatUpdate,
-    DataDictionary,
-    DataDictionaryResponse,
-    DataRegistryDataset,
-    DatasetCleansedResponse,
-    DictionaryCellUpdate,
-    EmptyResponse,
-    ExternalDataSourcesSelection,
-    ExternalDataStore,
-    FileUploadResponse,
-    GetBusinessAnalysisResult,
-    LoadDatabaseRequest,
-    RunAnalysisResult,
-    RunChartsResult,
-    RunDatabaseAnalysisResult,
-    SupportedDataSourceTypes,
+from core.middleware import session_middleware
+from core.routers import (
+    chats_router,
+    database_router,
+    datasets_router,
+    dictionaries_router,
+    external_data_stores_router,
+    registry_router,
+    supported_types_router,
+    user_router,
 )
 
 logger = get_logger()
-
-MAX_EXCEL_ROWS = 50000  # Maximum rows to export to Excel to prevent memory issues
 
 # Backward-compatible exports for utils.rest_api/core.rest_api consumers.
 SessionState = core_middleware.SessionState
 _initialize_session = core_middleware._initialize_session
 get_database = core_middleware.get_database
-session_middleware = core_middleware.session_middleware
 session_store = core_middleware.session_store
-
-
-def detect_and_decode_csv(raw_bytes: bytes, filename: str) -> str:
-    """
-    Detect encoding and decode CSV content with UTF-8 BOM handling.
-
-    Raises ValueError with a descriptive message when decoding fails.
-    """
-    if raw_bytes.startswith(b"\xef\xbb\xbf"):
-        logger.info(f"Stripped UTF-8 BOM from '{filename}'")
-        raw_bytes = raw_bytes[3:]
-
-    if not raw_bytes:
-        raise ValueError(f"File '{filename}' is empty or could not be read.")
-
-    detection_result = chardet.detect(raw_bytes)
-    detected_encoding = detection_result.get("encoding")
-    confidence = detection_result.get("confidence") or 0
-
-    if detected_encoding:
-        logger.info(
-            f"Detected encoding for '{filename}': {detected_encoding} "
-            f"(confidence: {confidence:.2f})"
-        )
-
-    encodings_to_try: list[str] = []
-    if detected_encoding and confidence >= 0.7:
-        encodings_to_try.append(detected_encoding)
-    encodings_to_try.extend(["utf-8", "cp932", "shift_jis"])
-
-    tried: set[str] = set()
-    for encoding in encodings_to_try:
-        normalized_encoding = encoding.lower()
-        if normalized_encoding in tried:
-            continue
-        tried.add(normalized_encoding)
-        try:
-            return raw_bytes.decode(encoding)
-        except (UnicodeDecodeError, LookupError) as e:
-            logger.debug(f"Failed to decode '{filename}' with encoding {encoding}: {e}")
-
-    raise ValueError(
-        f"Unable to decode file '{filename}'. All encoding attempts failed."
-    )
-
-
-def detect_delimiter(content: str) -> str:
-    """Detect a CSV delimiter from the first few lines, defaulting to comma."""
-    lines = content.split("\n")[:5]
-    delimiters = {",": 0, ";": 0, "\t": 0, "|": 0}
-
-    for line in lines:
-        if not line.strip():
-            continue
-        for delimiter in delimiters:
-            delimiters[delimiter] += line.count(delimiter)
-
-    best_delimiter = max(delimiters, key=lambda delimiter: delimiters[delimiter])
-    return best_delimiter if delimiters[best_delimiter] > 0 else ","
-
-
-def load_and_validate_csv(decoded_content: str, filename: str) -> pd.DataFrame:
-    """
-    Parse CSV content into a pandas DataFrame and validate that rows are present.
-
-    Raises ValueError with a descriptive message on parse or validation failure.
-    """
-    if "\r" in decoded_content:
-        decoded_content = decoded_content.replace("\r\n", "\n").replace("\r", "\n")
-
-    separator = detect_delimiter(decoded_content)
-    if separator != ",":
-        logger.info(f"Detected delimiter '{separator}' for '{filename}'")
-
-    try:
-        dataframe = pd.read_csv(io.StringIO(decoded_content), sep=separator)
-    except pd.errors.EmptyDataError as e:
-        raise ValueError(f"CSV file '{filename}' is empty.") from e
-    except pd.errors.ParserError as e:
-        raise ValueError(f"Unable to parse CSV file '{filename}'. Error: {e}") from e
-
-    if dataframe.empty:
-        raise ValueError(f"CSV file '{filename}' contains only headers, no data rows.")
-
-    lines = [line for line in decoded_content.split("\n") if line.strip()]
-    if len(lines) > 1:
-        expected_rows = len(lines) - 1
-        if len(dataframe) < expected_rows * 0.9:
-            logger.warning(
-                f"CSV file '{filename}' had inconsistent row lengths. "
-                f"Expected ~{expected_rows} rows, got {len(dataframe)}."
-            )
-
-    return dataframe
-
-
-def _chart_trace_to_dataframe(trace: dict[str, Any]) -> pd.DataFrame:
-    sequence_columns: dict[str, list[Any]] = {}
-    scalar_columns: dict[str, Any] = {}
-    max_len = 0
-
-    def register_sequence(name: str, values: Iterable[Any]) -> None:
-        nonlocal max_len
-        # Convert to list for length and slicing
-        values_list = list(values)
-        if len(values_list) > MAX_EXCEL_ROWS:
-            logger.warning(
-                f"Sequence '{name}' has {len(values_list)} items, truncating to {MAX_EXCEL_ROWS}."
-            )
-            values_list = values_list[:MAX_EXCEL_ROWS]
-
-        # Use list comprehension for normalization
-        def normalize(item):
-            if isinstance(item, (dict, list, tuple)):
-                try:
-                    return json.dumps(item)
-                except Exception:
-                    return str(item)
-            else:
-                return item
-
-        normalized = [normalize(item) for item in values_list]
-        sequence_columns[name] = normalized
-        max_len = max(max_len, len(normalized))
-
-    def register_scalar(name: str, value: Any) -> None:
-        scalar_columns[name] = value
-
-    def walk(current_key: str, value: Any) -> None:
-        if isinstance(value, (list, tuple)):
-            register_sequence(current_key, value)
-        elif isinstance(value, dict):
-            for sub_key, sub_value in value.items():
-                nested_key = f"{current_key}_{sub_key}" if current_key else sub_key
-                walk(nested_key, sub_value)
-        else:
-            register_scalar(current_key, value)
-
-    for key, value in trace.items():
-        walk(key, value)
-
-    if max_len == 0:
-        max_len = 1
-
-    padded_columns: dict[str, list[Any]] = {}
-    for name, values in sequence_columns.items():
-        if len(values) < max_len:
-            padded_columns[name] = values + [None] * (max_len - len(values))
-        else:
-            padded_columns[name] = values
-
-    df = pd.DataFrame(padded_columns, index=range(max_len))
-    for name, value in scalar_columns.items():
-        df[name] = value
-
-    return df
-
-
-# Dependency to provide the initialized database
-async def get_initialized_db(request: Request) -> AnalystDB:
-    if (
-        not hasattr(request.state.session, "analyst_db")
-        or request.state.session.analyst_db is None
-    ):
-        if not request.headers.get("x-user-email"):
-            logger.error("x-user-email is required in order to initialize the database")
-        raise HTTPException(
-            status_code=400,
-            detail="Database not initialized.",
-        )
-
-    return cast(AnalystDB, request.state.session.analyst_db)
-
-
-script_name = os.environ.get("SCRIPT_NAME", "")
-router = APIRouter(prefix=f"{script_name}/api/v1")
+get_initialized_db = deps.get_initialized_db
+detect_and_decode_csv = file_utils.detect_and_decode_csv
+detect_delimiter = file_utils.detect_delimiter
+load_and_validate_csv = file_utils.load_and_validate_csv
+create_chat = chats_routes.create_chat
+create_chat_message = chats_routes.create_chat_message
+create_new_chat_message = chats_routes.create_new_chat_message
+delete_chat = chats_routes.delete_chat
+delete_chat_message = chats_routes.delete_chat_message
+get_chat = chats_routes.get_chat
+get_chat_message = chats_routes.get_chat_message
+get_chat_messages = chats_routes.get_chat_messages
+get_chats = chats_routes.get_chats
+run_complete_analysis_task = chats_routes.run_complete_analysis_task
+save_chat_messages = chats_routes.save_chat_messages
+update_chat = chats_routes.update_chat
+get_and_process_tables = database_routes.get_and_process_tables
+get_database_tables = database_routes.get_database_tables
+load_from_database = database_routes.load_from_database
+process_and_update = database_routes.process_and_update
+delete_datasets = datasets_routes.delete_datasets
+download_dataset = datasets_routes.download_dataset
+get_cleansed_dataset = datasets_routes.get_cleansed_dataset
+get_dataset_by_id = datasets_routes.get_dataset_by_id
+get_dataset_metadata = datasets_routes.get_dataset_metadata
+upload_files = datasets_routes.upload_files
+delete_dictionary = dictionaries_routes.delete_dictionary
+download_dictionary = dictionaries_routes.download_dictionary
+get_dictionaries = dictionaries_routes.get_dictionaries
+update_dictionary_cell = dictionaries_routes.update_dictionary_cell
+get_available_external_data_stores = (
+    external_data_store_routes.get_available_external_data_stores
+)
+get_supported_datasource_types = (
+    external_data_store_routes.get_supported_datasource_types
+)
+register_external_data_sources = external_data_store_routes.register_external_data_sources
+update_data_sources_for_data_store = (
+    external_data_store_routes.update_data_sources_for_data_store
+)
+get_registry_datasets = registry_routes.get_registry_datasets
+get_datarobot_account = user_routes.get_datarobot_account
+store_datarobot_account = user_routes.store_datarobot_account
 _app_singleton: FastAPI | None = None
 
 
-def _create_fastapi_app(
-    lifespan: Lifespan[FastAPI] | None = None,
-    title: str = "Data Analyst API",
-) -> FastAPI:
-    created_app = FastAPI(
-        title=title,
-        description="""
-        An intelligent API for data analysis that provides capabilities including:
-        - Dataset management (upload CSV/Excel files, connect to databases, access the Data Registry)
-        - Data cleansing and standardization
-        - Data dictionary creation and management
-        - Chat-based data analysis conversations
-        - Python code generation
-        - Chart creation
-        - Business insights generation
-
-        Available endpoint groups:
-        - /api/v1/registry: Access Data Registry datasets
-        - /api/v1/database: Database connection and table management
-        - /api/v1/datasets: Upload, retrieve, and manage datasets
-        - /api/v1/dictionaries: Manage data dictionaries
-        - /api/v1/chats: Create and manage chat conversations for data analysis
-
-        The API uses OpenAI's GPT models for intelligent analysis and response generation.
-        """,
-        version="1.0.0",
-        contact={"name": "API Support", "email": "support@example.com"},
-        license_info={
-            "name": "Apache 2.0",
-            "url": "https://www.apache.org/licenses/LICENSE-2.0.html",
-        },
-        lifespan=lifespan,
-        debug=True,  # Stack traces will be exposed for 500 responses
-    )
-
-    created_app.add_middleware(
-        CORSMiddleware,
-        allow_origins=["*"],
-        allow_credentials=True,
-        allow_methods=["*"],
-        allow_headers=["*"],
-    )
-
-    created_app.middleware("http")(session_middleware)
-    created_app.openapi = lambda: custom_openapi(created_app)  # type: ignore[method-assign]
-    created_app.include_router(router)
-
-    try:
-        from core.customize.rest_api import router as customize_router
-
-        created_app.include_router(customize_router, prefix=f"{script_name}/api/v1")
-        logger.info("Customize API endpoints mounted")
-    except ImportError as e:
-        logger.warning("Customize module was not found: %s", e)
-    except Exception as e:
-        logger.exception("Failed to load customize API endpoints: %s", e)
-
-    telemetry.log_application_start()
-    telemetry.instrument_fastapi_app(created_app)
-
-    return created_app
-
-
-# Add custom OpenAPI schema
 def custom_openapi(app: FastAPI) -> dict[str, Any]:
+    """Generate custom OpenAPI schema with security definitions."""
     if app.openapi_schema:
         return app.openapi_schema
 
@@ -401,1484 +136,94 @@ def create_app(
     return _create_fastapi_app(lifespan=lifespan, title=title)
 
 
-# Make this sync as the DR requests are synchronous, if async this would block.
-@router.get("/registry/datasets")
-def get_registry_datasets(
-    request: Request, remote: bool = False, limit: int = 100
-) -> list[DataRegistryDataset]:
-    """Return all registry datasets
-
-    Args:
-        request (Request): _description_
-        filter_downloadable (bool, optional): _description_. Defaults to False.
-        limit (int, optional): _description_. Defaults to 100.
+def _create_fastapi_app(
+    lifespan: Lifespan[FastAPI] | None = None,
+    title: str = "Data Analyst API",
+) -> FastAPI:
+    """
+    Create and configure the FastAPI application.
 
     Returns:
-        list[DataRegistryDataset]: _description_
+        FastAPI: The configured FastAPI application instance
     """
-    with use_user_token(request, allow_use_builder_token=True):
-        return list_registry_datasets(remote=remote, limit=limit)
-
-
-@router.get("/database/tables")
-async def get_database_tables() -> list[str]:
-    return await get_external_database().get_tables()
-
-
-async def process_and_update(
-    dataset_names: List[str],
-    analyst_db: AnalystDB,
-    datasource_type: DataSourceType,
-) -> None:
-    async for _ in process_data_and_update_state(
-        dataset_names, analyst_db, datasource_type
-    ):
-        pass
-
-
-@router.post("/datasets/upload")
-async def upload_files(
-    request: Request,
-    background_tasks: BackgroundTasks,
-    data_source: str | InternalDataSourceType | None = None,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-    files: List[UploadFile] | None = None,
-    registry_ids: str | None = Form(None),
-) -> list[FileUploadResponse]:
-    normalized_data_source: InternalDataSourceType = (
-        InternalDataSourceType.FILE
-        if data_source is None
-        else InternalDataSourceType(data_source)
-        if isinstance(data_source, str)
-        else data_source
-    )
-    dataset_names = []
-    response: list[FileUploadResponse] = []
-    if files:
-        for file in files:
-            try:
-                logger.info(f"Processing file: {file.filename}")
-                file_size = file.size or 0
-                contents = await file.read()
-                if file.filename is None:
-                    continue
-
-                file_extension = os.path.splitext(file.filename)[1].lower()
-                logger.info(f"File extension: {file_extension}")
-
-                if file_extension == ".csv":
-                    logger.info(f"Loading CSV: {file.filename}")
-                    log_memory()
-                    try:
-                        decoded_content = detect_and_decode_csv(contents, file.filename)
-                        df = load_and_validate_csv(decoded_content, file.filename)
-                    except ValueError as e:
-                        raise HTTPException(status_code=400, detail=str(e)) from e
-                    except Exception as e:
-                        raise HTTPException(
-                            status_code=400,
-                            detail=(
-                                f"Failed to read CSV file '{file.filename}': {str(e)}"
-                            ),
-                        ) from e
-                    log_memory()
-                    dataset_name = os.path.splitext(file.filename)[0]
-                    dataset = AnalystDataset(name=dataset_name, data=df)
-
-                    # Register dataset with the database
-                    reg_result = await analyst_db.register_dataset(
-                        dataset, InternalDataSourceType.FILE, file_size=file_size
-                    )
-                    if not reg_result["success"]:
-                        error_response: FileUploadResponse = {
-                            "filename": file.filename,
-                            "error": reg_result["msg"],
-                        }
-                        response.append(error_response)
-                        continue
-
-                    # Add to processing queue
-                    dataset_names.append(dataset.name)
-
-                    file_response: FileUploadResponse = {
-                        "filename": file.filename,
-                        "content_type": file.content_type,
-                        "size": len(contents),
-                        "dataset_name": dataset_name,
-                    }
-                    response.append(file_response)
-
-                    logger.info(
-                        f"Registering dataset: {dataset_name}, shape={df.shape}"
-                    )
-
-                elif file_extension in [".xlsx", ".xls"]:
-                    base_name = os.path.splitext(file.filename)[0]
-                    try:
-                        excel_dataset = pd.read_excel(
-                            io.BytesIO(contents), sheet_name=None
-                        )
-                    except Exception as e:
-                        raise HTTPException(
-                            status_code=400,
-                            detail=(
-                                f"Unable to read Excel file '{file.filename}'. "
-                                f"Error: {str(e)}"
-                            ),
-                        ) from e
-                    if isinstance(excel_dataset, dict):
-                        for sheet_name, data in excel_dataset.items():
-                            dataset_name = f"{base_name}_{sheet_name}"
-                            dataset = AnalystDataset(name=dataset_name, data=data)
-                            reg_result = await analyst_db.register_dataset(
-                                dataset,
-                                InternalDataSourceType.FILE,
-                                file_size=file_size,
-                            )
-                            if not reg_result["success"]:
-                                error_response: FileUploadResponse = {
-                                    "filename": file.filename,
-                                    "error": reg_result["msg"],
-                                }
-                                response.append(error_response)
-                                continue
-                            # Add to processing queue
-                            dataset_names.append(dataset.name)
-
-                            excel_sheet_response: FileUploadResponse = {
-                                "filename": file.filename,
-                                "content_type": file.content_type,
-                                "size": len(contents),
-                                "dataset_name": dataset_name,
-                            }
-                            response.append(excel_sheet_response)
-                    elif isinstance(excel_dataset, pd.DataFrame):
-                        dataset_name = base_name
-                        dataset = AnalystDataset(name=dataset_name, data=excel_dataset)
-                        reg_result = await analyst_db.register_dataset(
-                            dataset, InternalDataSourceType.FILE, file_size=file_size
-                        )
-                        if not reg_result["success"]:
-                            error_response: FileUploadResponse = {
-                                "filename": file.filename,
-                                "error": reg_result["msg"],
-                            }
-                            response.append(error_response)
-                            continue
-                        # Add to processing queue
-                        dataset_names.append(dataset.name)
-
-                        excel_file_response: FileUploadResponse = {
-                            "filename": file.filename,
-                            "content_type": file.content_type,
-                            "size": len(contents),
-                            "dataset_name": dataset_name,
-                        }
-                        response.append(excel_file_response)
-                else:
-                    raise ValueError(f"Unsupported file type: {file_extension}")
-
-            except Exception as e:
-                logger.error(
-                    f"Exception processing file {file.filename}: {e}", exc_info=True
-                )
-                error_response: FileUploadResponse = {
-                    "filename": file.filename or "unknown_file",
-                    "error": str(e),
-                }
-                response.append(error_response)
-
-    # Process the data in the background (cleansing and dictionary generation)
-    if dataset_names:
-        background_tasks.add_task(
-            process_and_update, dataset_names, analyst_db, InternalDataSourceType.FILE
-        )
-
-    if registry_ids:
-        id_list: list[str] = json.loads(registry_ids)
-        if id_list:
-            with use_user_token(request, allow_use_builder_token=True):
-                if normalized_data_source == InternalDataSourceType.REMOTE_REGISTRY:
-                    dataframes, tasks = await register_remote_registry_datasets(
-                        request, id_list, analyst_db
-                    )
-                else:
-                    dataframes = await load_registry_datasets(id_list, analyst_db)
-                    tasks = []
-                for func, args, kwargs in tasks:
-                    background_tasks.add_task(func, *args, **kwargs)
-                dataset_names = [
-                    dataset.name for dataset in dataframes if not dataset.error
-                ]
-                background_tasks.add_task(
-                    process_and_update,
-                    dataset_names,
-                    analyst_db,
-                    InternalDataSourceType.REMOTE_REGISTRY
-                    if normalized_data_source == InternalDataSourceType.REMOTE_REGISTRY
-                    else InternalDataSourceType.REGISTRY,
-                )
-                for dts in dataframes:
-                    dts_response: FileUploadResponse = {
-                        "dataset_name": dts.name,
-                        "error": dts.error,
-                    }
-                    response.append(dts_response)
-
-    return response
-
-
-@router.post("/database/select")
-async def load_from_database(
-    data: LoadDatabaseRequest,
-    background_tasks: BackgroundTasks,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-    sample_size: int = 1_000,
-) -> list[str]:
-    await asyncio.gather(
-        *[
-            analyst_db.register_dataset(
-                AnalystDataset(name=table_name),
-                InternalDataSourceType.DATABASE,
-            )
-            for table_name in data.table_names
-        ]
-    )
-
-    if data.table_names:
-        background_tasks.add_task(
-            get_and_process_tables,
-            data.table_names,
-            analyst_db,
-            sample_size,
-            data.schema_name,
-        )
-
-    return data.table_names
-
-
-async def get_and_process_tables(
-    table_names: list[str],
-    analyst_db: AnalystDB,
-    sample_size: int,
-    schema_name: str | None = None,
-) -> None:
-    dataset_names = await get_external_database(schema=schema_name).get_data(
-        *table_names,
-        analyst_db=analyst_db,
-        sample_size=sample_size,
-    )
-    await process_and_update(dataset_names, analyst_db, InternalDataSourceType.DATABASE)
-
-
-@router.get("/dictionaries")
-async def get_dictionaries(
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> list[DataDictionaryResponse]:
-    # Get datasets from the database
-    dataset_names = await analyst_db.list_analyst_datasets()
-
-    # Fetch dictionaries for each dataset
-    dictionaries = []
-    for name in dataset_names:
-        dictionary = await analyst_db.get_data_dictionary(name)
-        if dictionary and len(dictionary.column_descriptions):
-            dictionaries.append(cast(DataDictionaryResponse, dictionary))
-        else:
-            dictionaries.append(
-                DataDictionaryResponse(
-                    name=name, column_descriptions=[], in_progress=True
-                )
-            )
-
-    return dictionaries if dictionaries else []
-
-
-@router.get("/datasets/{dataset_id}")
-async def get_dataset_by_id(
-    dataset_id: str,
-    skip: int = 0,
-    limit: int = 1000,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> DatasetCleansedResponse:
-    """
-    Get a dataset by its ID with pagination support.
-
-    Args:
-        dataset_id: The unique identifier of the dataset
-        skip: Number of records to skip (for pagination)
-        limit: Maximum number of records to return (for pagination)
-
-    Returns:
-        A response containing the dataset (similar to cleansed dataset format)
-
-    Raises:
-        HTTPException: If the dataset doesn't exist or cannot be retrieved
-    """
-    try:
-        from core.analyst_db import DatasetType
-
-        df = await analyst_db.dataset_handler.get_dataframe(
-            dataset_id,
-            expected_type=DatasetType.ANALYST_RESULT_DATASET,
-            max_rows=None,
-        )
-
-        if skip > 0 or limit > 0:
-            df = df.iloc[skip : skip + limit]
-
-        metadata = await analyst_db.dataset_handler.get_dataset_metadata(dataset_id)
-
-        dataset = AnalystDataset(
-            name=metadata.original_name,
-            data=df,
-        )
-
-        # Return in the same format as cleansed dataset
-        return DatasetCleansedResponse(
-            dataset_name=metadata.original_name,
-            cleaning_report=None,  # No cleaning report for analyst datasets
-            dataset=dataset,
-        )
-    except ValueError as e:
-        raise HTTPException(
-            status_code=404, detail=f"Dataset '{dataset_id}' not found: {str(e)}"
-        )
-    except Exception as e:
-        logger.error(f"Error retrieving dataset '{dataset_id}': {str(e)}")
-        raise HTTPException(
-            status_code=500, detail=f"Error retrieving dataset: {str(e)}"
-        )
-
-
-@router.get("/datasets/{dataset_id}/download")
-async def download_dataset(
-    dataset_id: str,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-    bom: bool = False,
-) -> Response:
-    """
-    Download a dataset by ID as a CSV file.
-
-    Args:
-        dataset_id: The unique identifier of the dataset to download
-        bom: Whether to include UTF-8 BOM for Excel compatibility (default: False)
-
-    Returns:
-        CSV file attachment
-
-    Raises:
-        HTTPException: If the dataset doesn't exist or cannot be retrieved
-    """
-    try:
-        from core.analyst_db import DatasetType
-
-        df = await analyst_db.dataset_handler.get_dataframe(
-            dataset_id,
-            expected_type=DatasetType.ANALYST_RESULT_DATASET,
-            max_rows=None,
-        )
-
-        csv_content = io.StringIO()
-        df.write_csv(csv_content)
-
-        csv_text = csv_content.getvalue()
-        if bom:
-            csv_text = "\ufeff" + csv_text
-
-        response = Response(content=csv_text)
-        response.headers["Content-Type"] = "text/csv; charset=utf-8"
-
-        filename = f"analysis_result_dataset_{dataset_id[:8]}.csv"
-        response.headers["Content-Disposition"] = f'attachment; filename="{filename}"'
-
-        return response
-
-    except ValueError as e:
-        raise HTTPException(
-            status_code=404, detail=f"Dataset '{dataset_id}' not found: {str(e)}"
-        )
-    except Exception as e:
-        logger.error(f"Error downloading dataset '{dataset_id}': {str(e)}")
-        raise HTTPException(
-            status_code=500, detail=f"Error downloading dataset: {str(e)}"
-        )
-
-
-@router.get("/datasets/{name}/metadata")
-async def get_dataset_metadata(
-    name: str, analyst_db: AnalystDB = Depends(get_initialized_db)
-) -> DatasetMetadata:
-    """
-    Get metadata for a dataset by name from the database.
-
-    Args:
-        name: The name of the dataset
-
-    Returns:
-        A dictionary containing dataset metadata including type, creation date, columns, and row count
-
-    Raises:
-        HTTPException: If the dataset doesn't exist or cannot be retrieved
-    """
-    try:
-        return await analyst_db.get_dataset_metadata(name)
-    except ValueError as e:
-        raise HTTPException(
-            status_code=404, detail=f"Dataset metadata not found: {str(e)}"
-        )
-    except Exception as e:
-        raise HTTPException(
-            status_code=500, detail=f"Error retrieving dataset metadata: {str(e)}"
-        )
-
-
-@router.get("/datasets/{name}/cleansed")
-async def get_cleansed_dataset(
-    name: str,
-    skip: int = 0,
-    limit: int = 10000,
-    search: str | None = None,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> DatasetCleansedResponse:
-    """
-    Get a cleansed dataset by name from the database with pagination and search support.
-
-    Args:
-        name: The name of the dataset
-        skip: Number of records to skip (for pagination)
-        limit: Maximum number of records to return (for pagination)
-        search: Optional search term to filter columns by name (raw data view)
-
-    Returns:
-        A dictionary containing the cleaning report (if available) and the dataset as a list of records.
-
-    Raises:
-        HTTPException: If the dataset doesn't exist or cannot be retrieved
-    """
-    try:
-        ds_display = await analyst_db.get_dataset(name)
-
-        # Initialize the response structure
-        response = DatasetCleansedResponse(
-            dataset_name=name,
-            cleaning_report=None,
-            dataset=None,
-        )
-
-        # Attempt to fetch the cleaning report
-        try:
-            ds_display_cleansed = await analyst_db.get_cleansed_dataset(name)
-            response.cleaning_report = ds_display_cleansed.generate_cleaning_report()
-
-        except ValueError:
-            # Cleaning report is not available
-            response.cleaning_report = None
-
-        # Convert the dataset to a DataFrame
-        df_display = ds_display.to_df()
-
-        # Apply search filter if provided - filter columns by name
-        if search and search.strip():
-            search_term = search.strip().lower()
-            original_columns = list(df_display.columns)
-            matching_columns = [
-                col for col in original_columns if search_term in col.lower()
-            ]
-            if matching_columns:
-                # Select only the matching columns
-                try:
-                    df_display = df_display.loc[:, matching_columns]
-                except Exception as e:
-                    logger.error(f"Error selecting columns: {e}")
-                    # Fallback to original dataframe if selection fails
-                    pass
-            else:
-                # No matching columns, create empty dataframe
-                df_display = df_display.iloc[0:0, []]
-
-        # Apply pagination (skip and limit)
-        if skip > 0 or limit > 0:
-            df_display = df_display.iloc[skip : skip + limit]
-
-        # Create an instance of AnalystDataset
-        dataset = AnalystDataset(
-            name=name,
-            columns=df_display.columns,
-            data=df_display.to_dict(
-                orient="records"
-            ),  # Convert rows to a list of dictionaries
-        )
-
-        # Add the dataset to the response
-        response.dataset = dataset
-
-        return response
-
-    except ValueError as e:
-        raise HTTPException(
-            status_code=404, detail=f"Dataset '{name}' not found: {str(e)}"
-        )
-    except Exception as e:
-        logger.error(f"Error retrieving cleansed dataset '{name}': {str(e)}")
-        raise HTTPException(
-            status_code=500, detail=f"Error retrieving cleansed dataset: {str(e)}"
-        )
-
-
-@router.delete("/datasets", status_code=200)
-async def delete_datasets(
-    request: Request,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> None:
-    await analyst_db.delete_all_tables()
-
-
-@router.delete("/dictionaries/{name}", status_code=200)
-async def delete_dictionary(
-    name: str, analyst_db: AnalystDB = Depends(get_initialized_db)
-) -> None:
-    await analyst_db.delete_table(name)
-
-
-@router.get("/dictionaries/{name}/download")
-async def download_dictionary(
-    name: str,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-    bom: bool = False,
-) -> Response:
-    """
-    Download a dictionary as a CSV file.
-
-    Args:
-        name: Name of the dataset whose dictionary to download
-
-    Returns:
-        CSV file attachment
-    """
-    dictionary = await analyst_db.get_data_dictionary(name)
-
-    if not dictionary:
-        raise HTTPException(status_code=404, detail=f"Dictionary '{name}' not found")
-
-    # Convert the dictionary to a DataFrame
-    df = dictionary.to_application_df()
-
-    # Convert to CSV
-    logger.info(f"Converting dictionary to CSV for {name}")
-    csv_content = io.StringIO()
-    df.to_csv(csv_content, index=False)
-    logger.info(f"CSV conversion complete for {name}")
-
-    # Create response with CSV attachment
-    csv_text = csv_content.getvalue()
-    if bom:
-        csv_text = (
-            "\ufeff" + csv_text
-        )  # Prefix UTF-8 BOM for Excel compatibility with international characters
-    response = Response(content=csv_text)
-    response.headers["Content-Type"] = "text/csv; charset=utf-8"
-
-    return response
-
-
-@router.patch("/dictionaries/{name}/cells")
-async def update_dictionary_cell(
-    name: str,
-    update: DictionaryCellUpdate,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> DataDictionary:
-    dictionary = await analyst_db.get_data_dictionary(name)
-
-    if not dictionary:
-        raise HTTPException(status_code=404, detail=f"Dictionary '{name}' not found")
-
-    if update.rowIndex < 0 or update.rowIndex >= len(dictionary.column_descriptions):
-        raise HTTPException(
-            status_code=400, detail=f"Row index {update.rowIndex} is out of range"
-        )
-
-    column_description = dictionary.column_descriptions[update.rowIndex]
-    if hasattr(column_description, update.field):
-        setattr(column_description, update.field, update.value)
-    else:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Field '{update.field}' not found in dictionary row",
-        )
-
-    await analyst_db.register_data_dictionary(dictionary, clobber=True)
-
-    return dictionary
-
-
-@router.post("/chats")
-async def create_chat(
-    chat: ChatCreate,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> dict[str, str]:
-    """Create a new chat with optional data source"""
-
-    chat_id = await analyst_db.create_chat(
-        chat_name=chat.name,
-        data_source=chat.data_source,
-    )
-
-    return {"id": chat_id}
-
-
-@router.get("/chats")
-async def get_chats(
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> list[dict[str, Any]]:
-    """Get all chats"""
-    chat_list = await analyst_db.get_chat_list()
-
-    return [
-        {
-            "id": chat["id"],
-            "name": chat["name"],
-            "data_source": chat.get("data_source", "catalog"),
-            "created_at": chat["created_at"],
-        }
-        for chat in chat_list
-    ]
-
-
-@router.get("/chats/{chat_id}")
-async def get_chat(
-    chat_id: str, analyst_db: AnalystDB = Depends(get_initialized_db)
-) -> ChatResponse:
-    """Get a specific chat by ID"""
-    chat = await analyst_db.get_chat_messages(chat_id=chat_id)
-
-    return {
-        "id": chat_id,
-        "messages": chat,
-    }
-
-
-@router.put("/chats/{chat_id}")
-async def update_chat(
-    chat_id: str, chat: ChatUpdate, analyst_db: AnalystDB = Depends(get_initialized_db)
-) -> dict[str, str]:
-    """Update a chat's name and/or data source"""
-    response_messages = []
-
-    # Update chat name if provided
-    if chat.name:
-        await analyst_db.rename_chat(chat_id, chat.name)
-        response_messages.append("renamed")
-
-    # Update data source if provided
-    if chat.data_source:
-        await analyst_db.update_chat_data_source(chat_id, chat.data_source)
-        response_messages.append("updated data source")
-
-    if not response_messages:
-        return {"message": f"No changes made to chat with ID {chat_id}"}
-
-    return {
-        "message": f"Chat with ID {chat_id} was {' and '.join(response_messages)} successfully"
-    }
-
-
-@router.delete("/chats/{chat_id}", status_code=200)
-async def delete_chat(
-    chat_id: str, analyst_db: AnalystDB = Depends(get_initialized_db)
-) -> dict[str, str]:
-    """Delete a chat"""
-    # Get all messages to clean up datasets
-    messages = await analyst_db.get_chat_messages(chat_id=chat_id)
-
-    # Clean up datasets for all messages
-    for message in messages:
-        await cleanup_message_datasets(analyst_db, message)
-
-    # Delete the chat
-    await analyst_db.delete_chat(chat_id=chat_id)
-
-    return {"message": f"Chat with ID {chat_id} deleted successfully"}
-
-
-@router.get("/chats/{chat_id}/messages")
-async def get_chat_messages(
-    chat_id: str,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> list[AnalystChatMessage]:
-    """Get messages for a specific chat"""
-
-    chat = await analyst_db.get_chat_messages(chat_id=chat_id)
-
-    return chat
-
-
-@router.delete("/chats/messages/{message_id}")
-async def delete_chat_message(
-    request: Request,
-    message_id: str,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> list[AnalystChatMessage]:
-    """Delete a specific message"""
-    try:
-        message = await analyst_db.get_chat_message(message_id=message_id)
-        if not message:
-            raise HTTPException(
-                status_code=404, detail=f"Message with ID {message_id} not found"
-            )
-        else:
-            # Clean up associated datasets before deleting the message
-            await cleanup_message_datasets(analyst_db, message)
-
-            await analyst_db.delete_chat_message(message_id=message_id)
-            messages = await analyst_db.get_chat_messages(
-                chat_id=message.chat_id,
-            )
-            return cast(list[AnalystChatMessage], list(messages))
-    except Exception as e:
-        logger.error(f"Error deleting message: {str(e)}")
-
-        return cast(list[AnalystChatMessage], [])
-
-
-@router.get("/chats/{chat_id}/messages/{message_id}")
-async def get_chat_message(
-    chat_id: str,
-    message_id: str,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> AnalystChatMessage:
-    """Get a specific message by ID from a specific chat"""
-    try:
-        message = await analyst_db.get_chat_message(message_id=message_id)
-        if not message:
-            raise HTTPException(
-                status_code=404, detail=f"Message with ID {message_id} not found"
-            )
-
-        # Verify the message belongs to the specified chat
-        if message.chat_id != chat_id:
-            raise HTTPException(
-                status_code=404,
-                detail=f"Message with ID {message_id} not found in chat {chat_id}",
-            )
-
-        return message
-    except Exception as e:
-        logger.error(f"Error getting message: {str(e)}")
-        raise HTTPException(
-            status_code=500, detail=f"Error retrieving message: {str(e)}"
-        )
-
-
-@router.post("/chats/messages")
-async def create_new_chat_message(
-    request: Request,
-    payload: ChatMessagePayload,
-    background_tasks: BackgroundTasks,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> dict[str, Union[str, list[AnalystChatMessage], None]]:
-    """Create a new chat and post a message to it"""
-
-    # Create a new chat
-    chat_id = await analyst_db.create_chat(
-        chat_name=payload.chatName,
-        data_source=payload.data_source,
-    )
-
-    # Create the user message
-    user_message = AnalystChatMessage(
-        role="user", content=payload.message, components=[]
-    )
-
-    message_id = await analyst_db.add_chat_message(
-        chat_id=chat_id,
-        message=user_message,
-    )
-
-    # Create valid messages for the chat request
-    valid_messages: list[ChatCompletionMessageParam] = [
-        user_message.to_openai_message_param()
-    ]
-
-    # Add the current message
-    valid_messages.append(
-        ChatCompletionUserMessageParam(role="user", content=payload.message)
-    )
-
-    # Create the chat request
-    chat_request = ChatRequest(messages=valid_messages)
-
-    # Run the analysis in the background
-    # Running the sync version whill get run in a different thread as the async version,
-    # despite the name, blocks for significant period. (When that gets fixed, we can change this.)
-    background_tasks.add_task(
-        run_complete_analysis_task,
-        chat_request,
-        payload.data_source,
-        analyst_db,
-        chat_id,
-        message_id,
-        payload.enable_chart_generation,
-        payload.enable_business_insights,
-        request,
-    )
-
-    chat_list = await analyst_db.get_chat_list()
-    chat_name = next((n["name"] for n in chat_list if n["id"] == chat_id), None)
-    messages = await analyst_db.get_chat_messages(chat_id=chat_id)
-
-    chat = {
-        "id": chat_id,
-        "name": chat_name,
-        "messages": messages,
-    }
-
-    return chat
-
-
-@router.post("/chats/{chat_id}/messages")
-async def create_chat_message(
-    request: Request,
-    chat_id: str,
-    payload: ChatMessagePayload,
-    background_tasks: BackgroundTasks,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> dict[str, Union[str, list[AnalystChatMessage], None]]:
-    """Post a message to a specific chat"""
-    messages = await analyst_db.get_chat_messages(chat_id=chat_id)
-
-    # Check if any message is in progress
-    in_progress = any(message.in_progress for message in messages)
-
-    # Check if cancelled
-    if not in_progress:
-        # Check if there's an existing summary (find last system message)
-        last_summary_idx = None
-        for i in range(len(messages) - 1, -1, -1):  # Iterate backwards through indices
-            if messages[i].role == "system":
-                last_summary_idx = i
-                break
-
-        # Build context: if summary exists, use [summary] + messages_after, else all messages
-        if last_summary_idx is not None:
-            context_messages = messages[last_summary_idx:]
-            logger.info(
-                f"[chat_id={chat_id}] Using summary + {len(context_messages) - 1} messages after"
-            )
-        else:
-            context_messages = messages
-
-        # Create valid messages for the chat request
-        valid_messages: list[ChatCompletionMessageParam] = [
-            msg.to_openai_message_param()
-            for msg in context_messages
-            if msg.content.strip()
-        ]
-
-        # Add the current message
-        valid_messages.append(
-            ChatCompletionUserMessageParam(role="user", content=payload.message)
-        )
-
-        # Check context usage
-        tokens_used = count_messages_tokens(valid_messages, ALTERNATIVE_LLM_BIG)
-        usage_pct = (tokens_used / MODEL_CONTEXT_WINDOW) * 100
-
-        logger.info(
-            f"[chat_id={chat_id}] Context: {tokens_used:,}/{MODEL_CONTEXT_WINDOW:,} tokens ({usage_pct:.1f}%)"
-        )
-
-        # Create and store the user message first (before summarization)
-        user_message = AnalystChatMessage(
-            role="user", content=payload.message, components=[]
-        )
-        message_id = await analyst_db.add_chat_message(
-            chat_id=chat_id, message=user_message
-        )
-        user_message.id = message_id
-
-        # Trigger summarization if over threshold
-        if tokens_used >= CONTEXT_WARNING_THRESHOLD:
-            logger.warning(
-                f"[chat_id={chat_id}] ⚠️  Context at {usage_pct:.1f}% - creating summary"
-            )
-
-            # Create new system message with in_progress=True
-            summary_message = AnalystChatMessage(
-                role="system",
-                content="Summarizing conversation...",
-                components=[],
-                in_progress=True,
-            )
-            summary_message.id = await analyst_db.add_chat_message(
-                chat_id=chat_id, message=summary_message
-            )
-
-            # Determine what to summarize: from last system message (if exists) to now, or all messages
-            if last_summary_idx is not None:
-                # Summarize from last system message onwards
-                messages_to_summarize = [
-                    msg.to_openai_message_param()
-                    for msg in messages[last_summary_idx:]
-                    if msg.content.strip()
-                ]
-            else:
-                # Summarize all messages
-                messages_to_summarize = [
-                    msg.to_openai_message_param()
-                    for msg in messages
-                    if msg.content.strip()
-                ]
-
-            # Create token tracker for summarization
-            summarization_tracker = TokenUsageTracker(
-                strategy=HeuristicTokenCountingStrategy()
-            )
-
-            try:
-                summary_text = await summarize_conversation(
-                    messages_to_summarize, token_tracker=summarization_tracker
-                )
-
-                logger.info(
-                    f"[chat_id={chat_id}] Summarization token usage: "
-                    f"{summarization_tracker.prompt_tokens} prompt + "
-                    f"{summarization_tracker.completion_tokens} completion = "
-                    f"{summarization_tracker.total_tokens} total tokens"
-                )
-
-                # Update the summary message with actual content
-                summary_message.content = summary_text
-                summary_message.in_progress = False
-
-                await analyst_db.update_chat_message(
-                    message_id=summary_message.id,
-                    message=summary_message,
-                )
-
-                logger.info(
-                    f"[chat_id={chat_id}] Summary stored ({len(summary_text)} chars)"
-                )
-
-                # Rebuild context using the new summary
-                summary_param = summary_message.to_openai_message_param()
-                valid_messages = [
-                    summary_param,
-                    valid_messages[-1],
-                ]  # summary + current message
-
-            except Exception as e:
-                logger.error(f"[chat_id={chat_id}] Failed to create summary: {e}")
-                # Mark summary as failed
-                summary_message.content = "Failed to create summary"
-                summary_message.in_progress = False
-                summary_message.error = str(e)
-                await analyst_db.update_chat_message(
-                    message_id=summary_message.id,
-                    message=summary_message,
-                )
-                # Continue without summary
-
-        # Create the chat request
-        chat_request = ChatRequest(messages=valid_messages)
-
-        # Run the analysis in the background
-        # Running the sync version whill get run in a different thread as the async version,
-        # despite the name, blocks for significant period. (When that gets fixed, we can change this.)
-        background_tasks.add_task(
-            run_complete_analysis_task,
-            chat_request,
-            payload.data_source,
-            analyst_db,
-            chat_id,
-            message_id,
-            payload.enable_chart_generation,
-            payload.enable_business_insights,
-            request,
-        )
-
-    chat_list = await analyst_db.get_chat_list()
-    chat_name = next((n["name"] for n in chat_list if n["id"] == chat_id), None)
-    messages = await analyst_db.get_chat_messages(chat_id=chat_id)
-
-    chat = {
-        "id": chat_id,
-        "name": chat_name,
-        "messages": messages,
-    }
-
-    return chat
-
-
-@router.get("/chats/{chat_id}/messages/download/")
-async def save_chat_messages(
-    request: Request,
-    chat_id: str,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-    message_id: str | None = None,
-) -> StreamingResponse:
-    """
-    This API controller saves a chat ID to an excel spreadsheet which
-    saves key information, then is streamed back to the user.
-    """
-    temp_files: list[str] = []
-
-    chat_messages: List[AnalystChatMessage] = await analyst_db.get_chat_messages(
-        chat_id=chat_id
-    )
-
-    # If a specific message_id is provided, filter messages to include
-    # only that user message and its following assistant response.
-    if message_id:
-        idx = next((i for i, m in enumerate(chat_messages) if m.id == message_id), None)
-        if idx is None or chat_messages[idx].role != "user":
-            raise HTTPException(detail="User message not found", status_code=404)
-
-        # Find the following assistant message
-        assistant_message = None
-        for i in range(idx + 1, len(chat_messages)):
-            if chat_messages[i].role == "assistant":
-                assistant_message = chat_messages[i]
-                break
-
-        # Include user message and assistant response if found
-        if assistant_message:
-            filtered_messages = [chat_messages[idx], assistant_message]
-        else:
-            filtered_messages = [chat_messages[idx]]
-
-        chat_messages = filtered_messages
-    if not chat_messages:
-        # Create an empty workbook with basic structure for empty chats
-        analysis_workbook = Workbook()
-        # Remove the default sheet
-        analysis_workbook.remove(analysis_workbook.active)
-
-        # Create a single sheet indicating the chat is empty
-        empty_sheet = analysis_workbook.create_sheet("Empty Chat")
-        empty_sheet["A1"] = "Chat Export"
-        empty_sheet["A3"] = "This chat contains no messages to export yet."
-
-        output = io.BytesIO()
-        analysis_workbook.save(output)
-        output.seek(0)
-
-        return StreamingResponse(
-            output,
-            media_type="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-        )
-
-    if any(msg.in_progress for msg in chat_messages):
-        raise HTTPException(
-            detail="Cannot download while a chat is in progress.", status_code=425
-        )
-    analysis_workbook = Workbook()
-    data_sheets_count = 0
-    charts_sheets_count = 0
-    report_sheets_count = 0
-
-    # Remove the initial default sheet created by Workbook()
-    analysis_workbook.remove(analysis_workbook.active)
-
-    for i, chat_message in enumerate(chat_messages):
-        # Skip system messages (summarization messages)
-        if chat_message.role == "system":
-            continue
-
-        if chat_message.role == "assistant":
-            # Handle Analysis Report sheet
-            report_sheets_count += 1
-            report_sheet_name = (
-                "Sheet" if report_sheets_count == 1 else f"Sheet {report_sheets_count}"
-            )
-            report_sheet = analysis_workbook.create_sheet(report_sheet_name)
-
-            report_sheet["A1"] = "Analysis Report"
-
-            # Find the previous user message by searching backwards
-            user_question = "No question found"
-            for j in range(i - 1, -1, -1):
-                if chat_messages[j].role == "user":
-                    user_question = chat_messages[j].content
-                    break
-
-            report_sheet["A3"] = "Question"
-            report_sheet["A4"] = user_question
-            report_sheet["A6"] = "Answer"
-            report_sheet["A7"] = chat_message.content
-
-            business_components: list[GetBusinessAnalysisResult] = [
-                component
-                for component in chat_message.components
-                if isinstance(component, GetBusinessAnalysisResult)
-            ]
-            for index, business_component in enumerate(business_components):
-                cell_index = index + 1  # Excel uses 1 indexing
-                report_sheet.cell(9, cell_index).value = "Bottom Line"
-                report_sheet.cell(10, cell_index).value = business_component.bottom_line
-
-                report_sheet.cell(12, cell_index).value = "Additional Insights"
-                report_sheet.cell(
-                    13, cell_index
-                ).value = business_component.additional_insights
-
-                report_sheet.cell(15, cell_index).value = "Follow-up Questions:"
-                for q_index, followup_question in enumerate(
-                    business_component.follow_up_questions
-                ):
-                    report_sheet.cell(
-                        16 + q_index, cell_index
-                    ).value = followup_question
-
-            # Handle Data sheets
-            run_analysis_components: List[
-                RunAnalysisResult | RunDatabaseAnalysisResult
-            ] = [
-                component
-                for component in chat_message.components
-                if isinstance(component, (RunAnalysisResult, RunDatabaseAnalysisResult))
-            ]
-            for run_analysis_component in run_analysis_components:
-                # Load dataset from storage (dataset field is excluded from serialization)
-                if not run_analysis_component.dataset_id:
-                    continue
-
-                try:
-                    from core.analyst_db import DatasetType
-
-                    df = await analyst_db.dataset_handler.get_dataframe(
-                        run_analysis_component.dataset_id,
-                        expected_type=DatasetType.ANALYST_RESULT_DATASET,
-                        max_rows=None,
-                    )
-                    metadata = await analyst_db.dataset_handler.get_dataset_metadata(
-                        run_analysis_component.dataset_id
-                    )
-                    dataset_to_export = AnalystDataset(
-                        name=metadata.original_name,
-                        data=df,
-                    )
-                except Exception as e:
-                    logger.error(
-                        f"Failed to load dataset {run_analysis_component.dataset_id}: {e}"
-                    )
-                    continue
-
-                data_sheets_count += 1
-                data_sheet_name = (
-                    "Data" if data_sheets_count == 1 else f"Data {data_sheets_count}"
-                )
-                data_sheet = analysis_workbook.create_sheet(data_sheet_name)
-
-                try:
-                    dataset: pd.DataFrame = dataset_to_export.data.df
-                    # Convert to pandas with error handling for large datasets
-                    pandas_df = dataset
-
-                    # Add size check to prevent memory issues and Excel limits
-                    original_rows = pandas_df.shape[0]
-                    if original_rows > MAX_EXCEL_ROWS:
-                        logger.warning(
-                            f"Dataset too large ({original_rows} rows), truncating to {MAX_EXCEL_ROWS} rows"
-                        )
-                        pandas_df = pandas_df.head(MAX_EXCEL_ROWS)
-                        # Add a notice row at the top of the sheet
-                        data_sheet.append(
-                            [
-                                f"NOTICE: Dataset truncated from {original_rows:,} to {MAX_EXCEL_ROWS:,} rows due to Excel limitations"
-                            ]
-                        )
-                        data_sheet.append([])
-                    for r in dataframe_to_rows(pandas_df, index=False, header=True):
-                        data_sheet.append(r)
-
-                except Exception as e:
-                    logger.error(f"Failed to process dataset: {e}")
-                    # Create error sheet instead of crashing
-                    data_sheet["A1"] = "Dataset Processing Error"
-                    data_sheet["A2"] = f"Error: {str(e)}"
-
-            # Handle Charts sheets
-            run_charts_components: List[RunChartsResult] = [
-                component
-                for component in chat_message.components
-                if isinstance(component, RunChartsResult)
-            ]
-            for run_chart_component in run_charts_components:
-                for f, js in [
-                    (run_chart_component.fig1, run_chart_component.fig1_json),
-                    (run_chart_component.fig2, run_chart_component.fig2_json),
-                ]:
-                    if not js or not f:
-                        continue
-                    charts_sheets_count += 1
-                    charts_sheet_name = f"Chart {charts_sheets_count}"
-                    charts_sheet = analysis_workbook.create_sheet(charts_sheet_name)
-
-                    # Save the chart data by removing the some keys from this field
-                    # We don't know what exactly is going to be graphed, so this is an easy way of covering
-                    # ourselves
-                    try:
-                        parsed_json = json.loads(js)
-                        data_list = parsed_json.get("data", [])
-                        if data_list and len(data_list) > 0:
-                            fig_json = data_list[
-                                0
-                            ].copy()  # Create copy to avoid modifying original
-                            [fig_json.pop(k, None) for k in ["marker", "name", "type"]]
-                            df = _chart_trace_to_dataframe(fig_json)
-                            for r in dataframe_to_rows(df, index=False, header=True):
-                                charts_sheet.append(r)
-
-                        with tempfile.NamedTemporaryFile(
-                            suffix=".png", delete=False
-                        ) as tmpfile:
-                            f.write_image(tmpfile.name)
-                            img = XLImage(tmpfile.name)
-                            charts_sheet.add_image(img, "F3")
-                            temp_files.append(tmpfile.name)
-                    except (
-                        json.JSONDecodeError,
-                        KeyError,
-                        IndexError,
-                        ValueError,
-                    ) as e:
-                        logger.warning(f"Failed to process chart data: {e}")
-                        # Create error sheet instead of crashing
-                        charts_sheet["A1"] = "Chart Processing Error"
-                        charts_sheet["A2"] = f"Error: {str(e)}"
-                    except Exception as e:
-                        logger.error(f"Unexpected error processing chart: {e}")
-                        continue  # Skip this chart but continue processing
-    output = io.BytesIO()
-    analysis_workbook.save(output)
-    output.seek(0)
-
-    # Create background task to cleanup temporary files
-    def cleanup_files(file_paths: List[str]) -> None:
-        for fp in file_paths:
-            if os.path.exists(fp):
-                os.remove(fp)
-
-    background_task = BackgroundTask(cleanup_files, temp_files) if temp_files else None
-
-    return StreamingResponse(
-        output,
-        media_type="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-        background=background_task,
-    )
-
-
-async def run_complete_analysis_task(
-    chat_request: ChatRequest,
-    data_source: str,
-    analyst_db: AnalystDB,
-    chat_id: str,
-    message_id: str,
-    enable_chart_generation: bool,
-    enable_business_insights: bool,
-    request: Request,
-) -> None:
-    """Run the complete analysis pipeline"""
-    source = get_data_source_type(data_source)
-    logger.debug(
-        "Running analysis for user.",
-        extra={
-            "data_source": data_source,
-            "user_id": analyst_db.user_id,
+    app = FastAPI(
+        title=title,
+        description="""
+        An intelligent API for data analysis that provides capabilities including:
+        - Dataset management (upload CSV/Excel files, connect to databases, access the Data Registry)
+        - Data cleansing and standardization
+        - Data dictionary creation and management
+        - Chat-based data analysis conversations
+        - Python code generation
+        - Chart creation
+        - Business insights generation
+
+        Available endpoint groups:
+        - /api/v1/registry: Access Data Registry datasets
+        - /api/v1/database: Database connection and table management
+        - /api/v1/datasets: Upload, retrieve, and manage datasets
+        - /api/v1/dictionaries: Manage data dictionaries
+        - /api/v1/chats: Create and manage chat conversations for data analysis
+
+        The API uses OpenAI's GPT models for intelligent analysis and response generation.
+        """,
+        version="1.0.0",
+        contact={"name": "API Support", "email": "support@example.com"},
+        license_info={
+            "name": "Apache 2.0",
+            "url": "https://www.apache.org/licenses/LICENSE-2.0.html",
         },
-    )
-    dataset_metadata = []
-    if source in [InternalDataSourceType.REGISTRY, InternalDataSourceType.FILE]:
-        dataset_metadata = (
-            await analyst_db.list_analyst_dataset_metadata(
-                InternalDataSourceType.REGISTRY
-            )
-        ) + (
-            await analyst_db.list_analyst_dataset_metadata(InternalDataSourceType.FILE)
-        )
-    else:
-        dataset_metadata = await analyst_db.list_analyst_dataset_metadata(source)
-
-    user_email = request.headers.get("x-user-email")
-    logger.info(f"Sending request on behalf of {user_email}")
-    telemetry_json = {
-        "user_email": user_email,
-        "user_msg": chat_request.messages[-1]["content"],
-    }
-    run_analysis_iterator = run_complete_analysis(
-        chat_request=chat_request,
-        data_source=source,
-        dataset_metadata=dataset_metadata,
-        analyst_db=analyst_db,
-        chat_id=chat_id,
-        message_id=message_id,
-        enable_chart_generation=enable_chart_generation,
-        enable_business_insights=enable_business_insights,
-        telemetry_json=telemetry_json,
-        request=request,
+        lifespan=lifespan,
+        debug=True,  # Stack traces will be exposed for 500 responses
     )
 
-    async for message in run_analysis_iterator:
-        if isinstance(message, AnalysisGenerationError):
-            break
-        else:
-            pass
-
-
-# Sync as this is a long blocking request
-@router.get("/available-external-data-stores")
-def get_available_external_data_stores(
-    request: Request, analyst_db: AnalystDB = Depends(get_initialized_db)
-) -> list[ExternalDataStore]:
-    """List all available datastores. (An available datastore
-    (a) has datasources configured and (b) has a supported driver.)
-
-    Args:
-        request (Request): HTTP request.
-        limit (int): Maximum entries to retrieve.
-        offset (int): Number of entries to skip.
-
-    Returns:
-        list[ExternalDataStore]: The given page of datastores.
-    """
-    with use_user_token(request, allow_use_builder_token=True):
-        return asyncio.run(
-            DataSourceRecipe.list_available_datastores(analyst_db.user_id)
-        )
-
-
-@router.put(
-    "/external-data-stores/{external_data_store_id}/external-data-sources/",
-    responses={404: {"model": ExceptionBody}},
-)
-async def register_external_data_sources(
-    request: Request,
-    selected_datasource_ids: ExternalDataSourcesSelection,
-    external_data_store_id: str,
-    background_tasks: BackgroundTasks,
-    analyst_db: AnalystDB = Depends(get_initialized_db),
-) -> EmptyResponse:
-    """
-    Add data sources for a datastore, registering them as datasets in the app.
-
-    Args:
-        request (Request): The request
-        selected_datasource_ids (ExternalDataSourcesSelection): Select data sources for a datasource
-        external_data_store_id (str): Select the data store
-        background_tasks (BackgroundTasks): Background tasks
-        analyst_db (AnalystDB, optional): The database. Defaults to Depends(get_initialized_db).
-
-    Returns:
-        ExternalDataStore: The updated data store
-    """
-    logger.debug(
-        "PUT /external-data-stores/%s/external-data-sources/", external_data_store_id
+    # Add CORS middleware
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origins=["*"],  # Allows all origins
+        allow_credentials=True,
+        allow_methods=["*"],  # Allows all methods
+        allow_headers=["*"],  # Allows all headers
     )
-    with use_user_token(request, allow_use_builder_token=True):
-        name = await DataSourceRecipe.get_canonical_name_for_datastore_id(
-            external_data_store_id
-        )
-        if not name:
-            raise HTTPException(
-                404, f"DataStore {external_data_store_id} does not exist."
-            )
 
-        background_tasks.add_task(
-            update_data_sources_for_data_store,
-            request,
-            selected_datasource_ids,
-            external_data_store_id,
-            analyst_db,
-        )
+    # Add session middleware
+    app.middleware("http")(session_middleware)
 
-        logger.debug(
-            "Registering data sources for data store %s (%s).",
-            name,
-            external_data_store_id,
-        )
-        _, tasks = await sync_data_sources_and_datasets(
-            request=request,
-            canonical_name=name,
-            analyst_db=analyst_db,
-            data_store_id=external_data_store_id,
-            selected_datasource_ids=selected_datasource_ids,
-        )
+    # Set custom OpenAPI schema
+    app.openapi = lambda: custom_openapi(app)  # type: ignore[method-assign]
 
-        logger.debug(
-            "Adding background tasks for data store %s.", external_data_store_id
-        )
-        for f, arg, kwargs in tasks:
-            background_tasks.add_task(f, *arg, **kwargs)
-        background_tasks.add_task(
-            process_and_update,
-            [d.path for d in selected_datasource_ids.selected_data_sources],
-            analyst_db,
-            ExternalDataStoreNameDataSourceType.from_name(name),
-        )
-    return EmptyResponse()
+    # Get script name from environment
+    script_name = os.environ.get("SCRIPT_NAME", "")
+    prefix = f"{script_name}/api/v1"
 
+    # Register all routers
+    app.include_router(registry_router, prefix=prefix)
+    app.include_router(database_router, prefix=prefix)
+    app.include_router(datasets_router, prefix=prefix)
+    app.include_router(dictionaries_router, prefix=prefix)
+    app.include_router(chats_router, prefix=prefix)
+    app.include_router(user_router, prefix=prefix)
+    app.include_router(external_data_stores_router, prefix=prefix)
+    app.include_router(supported_types_router, prefix=prefix)
 
-async def update_data_sources_for_data_store(
-    request: Request,
-    selected_datasource_ids: ExternalDataSourcesSelection,
-    external_data_store_id: str,
-    analyst_db: AnalystDB,
-) -> None:
-    with use_user_token(request, allow_use_builder_token=True):
-        logger.debug("Loading recipe for datasources %s.")
-        recipe = await DataSourceRecipe.load_or_create(
-            analyst_db, external_data_store_id
-        )
-        logger.debug("Updating recipe with data sources for %s", external_data_store_id)
-        await recipe.select_data_sources(selected_datasource_ids.selected_data_sources)
+    try:
+        from core.customize.rest_api import router as customize_router
 
+        app.include_router(customize_router, prefix=prefix)
+        logger.info("Customize API endpoints mounted")
+    except ImportError as e:
+        logger.warning("Customize module was not found: %s", e)
+    except Exception as e:
+        logger.exception("Failed to load customize API endpoints: %s", e)
 
-@router.get("/supported-data-source-types")
-async def get_supported_datasource_types(request: Request) -> SupportedDataSourceTypes:
-    types: list[InternalDataSourceType] = [
-        InternalDataSourceType.FILE,
-        InternalDataSourceType.REGISTRY,
-    ]
-    if not isinstance(get_external_database(), NoDatabaseOperator):
-        types.append(InternalDataSourceType.DATABASE)
-    if DatasetSparkRecipe.should_use_spark_recipe():
-        types.append(InternalDataSourceType.REMOTE_REGISTRY)
-    return SupportedDataSourceTypes(supported_types=[t.value for t in types])
+    # Initialize telemetry on application startup
+    telemetry.log_application_start()
 
+    # Setup auto-instrumentation for FastAPI
+    # This will automatically trace all incoming HTTP requests
+    telemetry.instrument_fastapi_app(app)
 
-@router.get("/user/datarobot-account")
-async def get_datarobot_account(
-    request: Request,
-) -> dict[str, Any]:
-    api_token = os.environ.get("DATAROBOT_API_TOKEN")
-    scoped_token = get_visitors_token(request)
-
-    truncated_api_token = None
-    if api_token:
-        truncated_api_token = f"****{api_token[-4:]}"
-
-    truncated_scoped_token = None
-    if scoped_token:
-        truncated_scoped_token = f"****{scoped_token[-4:]}"
-
-    return {
-        "datarobot_account_info": request.state.session.datarobot_account_info,
-        "datarobot_api_token": truncated_api_token,
-        "datarobot_api_scoped_token": truncated_scoped_token,
-    }
-
-
-@router.post("/user/datarobot-account")
-async def store_datarobot_account(
-    request: Request,
-) -> dict[str, Any]:
-    request_data = await request.json()
-
-    if "api_token" in request_data and request_data["api_token"]:
-        request.state.session.datarobot_api_scoped_token = request_data["api_token"]
-
-    return {"success": True}
+    return app
 
 
 app = create_app()
